requirements.txt
numpy==1.26.4
chromadb==0.5.0
sentence-transformers
streamlit
datasets
pillow
python-dotenv
langchain
langchain-openai
openai
tiktoken
